add multiple commands

+++ client.java
@@ -1,5 +1,6 @@
 import java.io.*;
 import java.net.*;
+import java.util.regex.*;
 
 class client{
 	
@@ -24,29 +25,33 @@ class client{
 			//tokenizing input string
 			String[] command_input = option.split("\\s+");
  
-			if(command_input[0].toUpperCase().equals("SEND")) {   //SEND command
+			if(command_input[0].toUpperCase().equals("SEND")) {   
 				//for uploading the file on server
 				ftp.sendfile(s, command_input);
 			}
-			else if(option.toUpperCase().equals("RECEIVE")) {
+			else if(command_input[0].toUpperCase().equals("RECEIVE")) {
 				//for downloading the file on server
-				System.out.println("RECEIVE command received");
-				ftp.receivefile(s);
+				ftp.receivefile(s, command_input);
 			}
-			else if(option.toUpperCase().equals("LIST")) {
+			else if(command_input[0].toUpperCase().equals("LIST")) {
 				//For listing available files on server
-				System.out.println("LIST command received");
 				ftp.listdirectory(s);
 			}
-			else if(option.toUpperCase().equals("CD")) {
+			else if(command_input[0].toUpperCase().equals("CD")) {
 				//for changing directory
-				System.out.println("CD command received");
-				ftp.cd(s);
+				ftp.cd(s, command_input);
 			}
-			else if(option.toUpperCase().equals("PWD")) {
+			else if(command_input[0].toUpperCase().equals("PWD")) {
 				//for getting current working directory
-				System.out.println("PWD command received");
 				ftp.pwd(s);
+			}
+			else if(command_input[0].toUpperCase().equals("HELP")) {
+				// Command for getting help
+				ftp.help();
+			}
+			else if(command_input[0].toUpperCase().equals("CLOSE")) {
+				// Command for getting help
+				ftp.close(s);
 			}
 		}
 	}
@@ -86,7 +91,7 @@ class client{
 		System.out.println("File Sent");
 	}
 	
-	public void receivefile(Socket s) throws Exception
+	public void receivefile(Socket s, String[] command_input) throws Exception
 	{
 		Socket ssock=s;
 		DataInputStream in=new DataInputStream(System.in);
@@ -95,22 +100,34 @@ class client{
 		
 		cout.writeUTF("SEND");
 		
-		String filename = in.readLine();
+		String filename = command_input[1];
 		cout.writeUTF(filename);
 		System.out.println("Receiving File " + filename);
 		
 		File f = new File(filename);
-		FileOutputStream fout = new FileOutputStream(f);
 		int ch;
+		String Response = cin.readUTF();
 		
-		
-		do {
-			ch=Integer.parseInt(cin.readUTF());
+		if(Response.equals("NOT EXIST")) {				//if file does not exist
+
+			System.out.println("File does not exists! Please Enter valid filename");
+			return;
+		}
+		else {
+
+			FileOutputStream fout = new FileOutputStream(f);
+			ch=Integer.parseInt(Response);
 			if(ch!=-1)
 				fout.write(ch);
-		}while(ch!=-1);
-		System.out.println("Received File...");
-		fout.close();
+			do {
+
+				ch=Integer.parseInt(cin.readUTF());
+				if(ch!=-1)
+					fout.write(ch);
+			}while(ch!=-1);
+			System.out.println("Received File...");
+			fout.close();
+		}
 	}
 	
 	public void listdirectory(Socket s) throws Exception {
@@ -121,10 +138,7 @@ class client{
 		DataOutputStream cout = new DataOutputStream(ssock.getOutputStream());
 
 		cout.writeUTF("LIST");
-		System.out.println("Response from the server:");
-
 		String data;
-
 		
 		while(true) {
 
@@ -135,7 +149,7 @@ class client{
 		}
 	}
 
-	public void cd(Socket s) throws Exception {
+	public void cd(Socket s, String[] command_input) throws Exception {
 
 		Socket ssock = s;
 
@@ -144,12 +158,13 @@ class client{
 		DataOutputStream cout = new DataOutputStream(ssock.getOutputStream());
 
 		cout.writeUTF("CD");
-		System.out.print("Enter path to directory: ");
-		String path = in.readLine();
+		String path = command_input[1];
 		cout.writeUTF(path);
 
 		if(cin.readUTF().equals("OK"))
 			System.out.println("Directory changed succefully");
+		else
+			System.out.println("Directory Not Found! Please Enter Valid Path.");
 	}
 
 	public void pwd(Socket s) throws Exception {
@@ -160,7 +175,39 @@ class client{
 		DataOutputStream cout = new DataOutputStream(ssock.getOutputStream());
 
 		cout.writeUTF("PWD");
-		System.out.println("Response from server:");
 		System.out.println(cin.readUTF());
+	}
+
+	public void close(Socket s) throws Exception {
+
+		Socket ssock = s;
+
+		DataInputStream cin = new DataInputStream(ssock.getInputStream());
+		DataOutputStream cout = new DataOutputStream(ssock.getOutputStream());
+
+		try {
+			
+			cout.writeUTF("CLOSE");
+			cin.close();
+			cout.close();
+			ssock.close();
+			System.out.println("Good Bye!");
+			System.exit(0);
+		}
+		catch  (SocketException ex) {
+        	System.out.println("Good Bye!");
+		}	
+	}
+
+	public void help() throws Exception {
+
+		System.out.println("\n\n============Commands=============");
+		System.out.println("\n\nsend filename -- To send file to server");
+		System.out.println("\n\nreceive filename -- To receive file from server");
+		System.out.println("\n\nlist -- To list all file available on current working directory");
+		System.out.println("\n\ncd path -- To change directory to desire path");
+		System.out.println("\n\npwd -- To know current working directory");
+		System.out.println("\n\nclose -- To exit from the program");
+		System.out.println("\n\nhelp -- To get help for commands");
 	}
 }
